requirements.txt
aiohttp==3.6.2
alembic==1.4.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
distlib==0.3.0
filelock==3.0.12
Flask==1.1.1
flask-heroku==0.1.9
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
multidict==4.7.5
newrelic==5.10.0.138
psycopg2==2.8.4
py-mini-racer==0.2.0
pyee==7.0.1
python-binary-memcached==0.29.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
six==1.14.0
slackclient==2.5.0
slackeventsapi==2.1.0
sq-native==0.5.1.0.1
SQLAlchemy==1.3.15
sqreen==1.19.0
uhashring==1.1
urllib3==1.25.8
virtualenv==20.0.15
Werkzeug==1.0.1
yarl==1.4.2
zipp==3.1.0
